Compete Day 4 - Lists, randomization and cinema booking system

+++ Day 03/logical_operators.py
@@ -1,7 +1,7 @@
 # modulo operator %
 # determine if the number is odd or even
 print("ODD OR EVEN ?")
-number = float(input("\nEnter a number: "))
+number = int(input("\nEnter a number: "))
 remainder = number % 2
 if remainder == 0:
     print(f"\nThe number {number} is an even number !")
